refactor(W3D2): domain, infrastructure로 패키지 분리
- domain (order, voucher)
- infrastructure (configuration, order, voucher)
