+++ kdt-spring-order/src/main/java/org/prgrms/kdt/domain/order/Order.java
@@ -1,4 +1,6 @@
-package org.prgrms.kdt;
+package org.prgrms.kdt.domain.order;
+
+import org.prgrms.kdt.domain.voucher.Voucher;
 
 import java.util.List;
 import java.util.Optional;
