+++ kdt-spring-order/src/main/java/org/prgrms/kdt/domain/order/OrderItem.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt;
+package org.prgrms.kdt.domain.order;
 
 import java.util.UUID;
