+++ kdt-spring-order/src/main/java/org/prgrms/kdt/domain/order/OrderRepository.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt;
+package org.prgrms.kdt.domain.order;
 
 public interface OrderRepository {
     void insert(Order order);
