+++ kdt-spring-order/src/main/java/org/prgrms/kdt/domain/order/OrderStatus.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt;
+package org.prgrms.kdt.domain.order;
 
 public enum OrderStatus {
     ACCEPTED,
